refactor: chanfe folder structure for login module
